bench request-vc
